lambda/s3_pdf_processor/requirements.txt
boto3
PyPDF2
psycopg2-binary
